requirements.txt
Django==1.8.6
djangorestframework==3.3.1
requests==2.8.1
wsgiref==0.1.2
